Fix genshify arg order, and only remove intermediate directory if it exists

+++ Makefile
@@ -49,11 +49,11 @@ out/%.txt: %.genshi.txt book_helpers.py $(OUT)
 
 out/%.html: tmp/%.html $(OUT) $(TMP)
 	# Remove temporary files to ensure regeneration of lilypond pictures, since lilypond-book does its own incomplete dependency check
-	rm -r out/lily/
+	if [ -d out/lily/ ] ; then rm -r out/lily/ ; fi
 	lilypond-book --lily-output-dir out/lily/ --process "lilypond -dbackend=eps" --output out/ $<
 
 tmp/%.html: %.lilypond-genshi.html book_helpers.py chromaturn.ly $(TMP)
-	./genshify $< -o $@ ${genshify_args}
+	./genshify -o $@ $< ${genshify_args}
 
 local: build_all
 	mkdir -p ~/frasergo-website/frasergo-mezzanine/static/projects/harmonihue/
